feat: database migrations for the data transfer process with three-row separation
