+++ .gitignore
@@ -12,3 +12,4 @@ __pycache__/
 # --- Virtual env ---
 venv/
 .env
+lost_in_ai_BE/.env
