+++ alembic/versions/e65f82af81bc_create_user_table.py
@@ -1,0 +1,39 @@
+"""create_user_table
+
+Revision ID: e65f82af81bc
+Revises: 
+Create Date: 2025-09-22 18:56:32.134666
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'e65f82af81bc'
+down_revision: Union[str, Sequence[str], None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    op.create_table(
+        "users",
+        sa.Column("id", sa.Integer(), primary_key=True),
+        sa.Column("username", sa.String(50), nullable=False),
+        sa.Column("email", sa.String(100), nullable=False),
+        sa.Column("name", sa.String(50), nullable=False),
+        sa.Column("surname", sa.String(50), nullable=False),
+        sa.Column("date_of_birth", sa.Date(), nullable=False),
+        sa.Column("created_at", sa.TIMESTAMP(timezone=False), nullable=False, server_default=sa.text("CURRENT_TIMESTAMP")),
+        sa.UniqueConstraint("username", name="uq_users_username"),
+        sa.UniqueConstraint("email", name="uq_users_email"),
+    )
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    op.drop_table("users")
